+++ packages/brick_offline_first/CHANGELOG.md
@@ -4,9 +4,11 @@
 * Optionally ignore Tunnel not found requests (these occur when connectivity exists but the queried endpoint is unreachable) when making repository requests
 * Adds argument to repository to reattempt requests based on the status code from the response
 * `OfflineRequestQueue#process` became a protected method
-* Added `OfflineRequestQueue#unprocessedJobs` that invokes new method `RequestSqliteCache.unprocessedJobs`.
-* Renamed `RequestSqliteCache.unprocessedRequests` to `RequestSqliteCache.latestUnprocessedRequest` as the expected query only returns one locked row at a time.
-* `RequestSqliteCache.latestUnprocessedRequest` locks _all_ unprocessed rows, not just the first one
+* Added `RequestSqliteCacheManager` to interact with the queue. This new class receives most static methods from `RequestSqliteCache`.
+* Added `OfflineRequestQueue#requestManager` to access queue via a `RequestSqliteCacheManager` instance.
+* Renamed `RequestSqliteCache.unprocessedRequests` to `RequestSqliteCacheManager.prepareNextRequestToProcess` as the expected query only returns one locked row at a time.
+* `RequestSqliteCacheManager.prepareNextRequestToProcess` locks _all_ unprocessed rows, not just the first one
+* Private member `OfflineFirstWithRestRepository#offlineRequestQueue` is now protected
 
 ## 0.0.5+1